feat(legacy view template): extract view to a separate file
